+++ alicloud/resource_alicloud_fc_trigger_test.go
@@ -2,9 +2,8 @@ package alicloud
 
 import (
 	"fmt"
+	"regexp"
 	"testing"
-
-	"strings"
 
 	"github.com/aliyun/aliyun-log-go-sdk"
 	"github.com/aliyun/fc-go-sdk"
@@ -51,6 +50,269 @@ func TestAccAlicloudFCTrigger_log(t *testing.T) {
 	})
 }
 
+func TestAccAlicloudFCTrigger_mnsTopic(t *testing.T) {
+	var service fc.GetServiceOutput
+	var project sls.LogProject
+	var store sls.LogStore
+	var function fc.GetFunctionOutput
+	var trigger fc.GetTriggerOutput
+
+	randInt := acctest.RandInt()
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheckWithRegions(t, false, connectivity.FcNoSupportedRegions) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckAlicloudFCTriggerDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAlicloudFCTriggerMnsTopic(testTriggerMnsTopicTemplate, testFCMnsTopicRoleTemplate, testFCMnsTopicPolicyTemplate, randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAlicloudLogProjectExists("alicloud_log_project.foo", &project),
+					testAccCheckAlicloudLogStoreExists("alicloud_log_store.foo", &store),
+					testAccCheckAlicloudFCServiceExists("alicloud_fc_service.foo", &service),
+					testAccCheckAlicloudFCFunctionExists("alicloud_fc_function.foo", &function),
+					testAccCheckAlicloudFCTriggerExists("alicloud_fc_trigger.foo", &trigger),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "name", fmt.Sprintf("tf-testacc-fc-trigger-mns-topic-%v", randInt)),
+					resource.TestCheckResourceAttrSet("alicloud_fc_trigger.foo", "config_mns"),
+
+					resource.TestMatchResourceAttr("alicloud_fc_trigger.foo", "source_arn", regexp.MustCompile(fmt.Sprintf("acs:mns:[a-z0-9-]+:[a-z0-9]+:/topics/tf-testacc-fc-trigger-mns-topic-%v", randInt))),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "config_mns", testTriggerMnsTopicTemplate),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "type", "mns_topic"),
+				),
+			},
+			{
+				Config: testAlicloudFCTriggerMnsTopicUpdate(testTriggerMnsTopicTemplateUpdate, testFCMnsTopicRoleTemplate, testFCMnsTopicPolicyTemplate, randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckAlicloudFCTriggerExists("alicloud_fc_trigger.foo", &trigger),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "name", fmt.Sprintf("tf-testacc-fc-trigger-mns-topic-%v", randInt)),
+					resource.TestCheckResourceAttrSet("alicloud_fc_trigger.foo", "config_mns"),
+
+					resource.TestMatchResourceAttr("alicloud_fc_trigger.foo", "source_arn", regexp.MustCompile(fmt.Sprintf("acs:mns:[a-z0-9-]+:[a-z0-9]+:/topics/tf-testacc-fc-trigger-mns-topic-%v", randInt))),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "config_mns", testTriggerMnsTopicTemplateUpdate),
+					resource.TestCheckResourceAttr("alicloud_fc_trigger.foo", "type", "mns_topic"),
+				),
+			},
+		},
+	})
+}
+
+func testAlicloudFCTriggerMnsTopic(trigger, role, policy string, randInt int) string {
+	return fmt.Sprintf(`
+variable "name" {
+  default = "tf-testacc-fc-trigger-mns-topic-%v"
+}
+data "alicloud_regions" "current_region" {
+  current = true
+}
+data "alicloud_account" "current" {
+}
+resource "alicloud_log_project" "foo" {
+  name = "${var.name}"
+  description = "tf unit test"
+}
+resource "alicloud_log_store" "bar" {
+  project = "${alicloud_log_project.foo.name}"
+  name = "${var.name}-source"
+  retention_period = "3000"
+  shard_count = 1
+}
+resource "alicloud_log_store" "foo" {
+  project = "${alicloud_log_project.foo.name}"
+  name = "${var.name}"
+  retention_period = "3000"
+  shard_count = 1
+}
+resource "alicloud_mns_topic" "foo" {
+  name = "${var.name}"
+}
+resource "alicloud_fc_service" "foo" {
+  name = "${var.name}"
+  internet_access = false
+}
+resource "alicloud_oss_bucket" "foo" {
+  bucket = "${var.name}"
+}
+resource "alicloud_oss_bucket_object" "foo" {
+  bucket = "${alicloud_oss_bucket.foo.id}"
+  key = "fc/hello.zip"
+  content = <<EOF
+  	# -*- coding: utf-8 -*-
+	def handler(event, context):
+	    print "hello world"
+	    return 'hello world'
+  EOF
+}
+resource "alicloud_fc_function" "foo" {
+  service = "${alicloud_fc_service.foo.name}"
+  name = "${var.name}"
+  oss_bucket = "${alicloud_oss_bucket.foo.id}"
+  oss_key = "${alicloud_oss_bucket_object.foo.key}"
+  memory_size = 512
+  runtime = "python2.7"
+  handler = "hello.handler"
+}
+resource "alicloud_ram_role" "foo" {
+  name = "${var.name}-trigger"
+  document = <<EOF
+  %s
+  EOF
+  description = "this is a test"
+  force = true
+}
+resource "alicloud_ram_policy" "foo" {
+  name = "${var.name}-trigger"
+  document = <<EOF
+  %s
+  EOF
+  description = "this is a test"
+  force = true
+}
+resource "alicloud_ram_role_policy_attachment" "foo" {
+  role_name = "${alicloud_ram_role.foo.name}"
+  policy_name = "${alicloud_ram_policy.foo.name}"
+  policy_type = "Custom"
+}
+resource "alicloud_fc_trigger" "foo" {
+  service = "${alicloud_fc_service.foo.name}"
+  function = "${alicloud_fc_function.foo.name}"
+  name = "${var.name}"
+  role = "${alicloud_ram_role.foo.arn}"
+  source_arn = "acs:mns:${data.alicloud_regions.current_region.regions.0.id}:${data.alicloud_account.current.id}:/topics/${alicloud_mns_topic.foo.name}"
+  type = "mns_topic"
+  config_mns = <<EOF
+  %s
+  EOF
+  depends_on = ["alicloud_ram_role_policy_attachment.foo"]
+}
+`, randInt, role, policy, trigger)
+}
+
+func testAlicloudFCTriggerMnsTopicUpdate(trigger, role, policy string, randInt int) string {
+	return fmt.Sprintf(`
+variable "name" {
+  default = "tf-testacc-fc-trigger-mns-topic-%v"
+}
+data "alicloud_regions" "current_region" {
+  current = true
+}
+data "alicloud_account" "current" {
+}
+resource "alicloud_log_project" "foo" {
+  name = "${var.name}"
+  description = "tf unit test"
+}
+resource "alicloud_log_store" "bar" {
+  project = "${alicloud_log_project.foo.name}"
+  name = "${var.name}-source"
+  retention_period = "3000"
+  shard_count = 1
+}
+resource "alicloud_log_store" "foo" {
+  project = "${alicloud_log_project.foo.name}"
+  name = "${var.name}"
+  retention_period = "3000"
+  shard_count = 1
+}
+resource "alicloud_mns_topic" "foo" {
+  name = "${var.name}"
+}
+resource "alicloud_fc_service" "foo" {
+  name = "${var.name}"
+  internet_access = false
+}
+resource "alicloud_oss_bucket" "foo" {
+  bucket = "${var.name}"
+}
+resource "alicloud_oss_bucket_object" "foo" {
+  bucket = "${alicloud_oss_bucket.foo.id}"
+  key = "fc/hello.zip"
+  content = <<EOF
+  	# -*- coding: utf-8 -*-
+	def handler(event, context):
+	    print "hello world"
+	    return 'hello world'
+  EOF
+}
+resource "alicloud_fc_function" "foo" {
+  service = "${alicloud_fc_service.foo.name}"
+  name = "${var.name}"
+  oss_bucket = "${alicloud_oss_bucket.foo.id}"
+  oss_key = "${alicloud_oss_bucket_object.foo.key}"
+  memory_size = 512
+  runtime = "python2.7"
+  handler = "hello.handler"
+}
+resource "alicloud_fc_trigger" "foo" {
+  service = "${alicloud_fc_service.foo.name}"
+  function = "${alicloud_fc_function.foo.name}"
+  name = "${var.name}"
+  role = "${alicloud_ram_role.foo.arn}"
+  source_arn = "acs:mns:${data.alicloud_regions.current_region.regions.0.id}:${data.alicloud_account.current.id}:/topics/${alicloud_mns_topic.foo.name}"
+  type = "mns_topic"
+  config_mns = <<EOF
+  %s
+  EOF
+  depends_on = ["alicloud_ram_role_policy_attachment.foo"]
+}
+resource "alicloud_ram_role" "foo" {
+  name = "${var.name}-trigger"
+  document = <<EOF
+  %s
+  EOF
+  description = "this is a test"
+  force = true
+}
+resource "alicloud_ram_policy" "foo" {
+  name = "${var.name}-trigger"
+  document = <<EOF
+  %s
+  EOF
+  description = "this is a test"
+  force = true
+}
+resource "alicloud_ram_role_policy_attachment" "foo" {
+  role_name = "${alicloud_ram_role.foo.name}"
+  policy_name = "${alicloud_ram_policy.foo.name}"
+  policy_type = "Custom"
+}
+`, randInt, trigger, role, policy)
+}
+
+var testTriggerMnsTopicTemplate = `{"filterTag":"testTag","notifyContentFormat":"STREAM","notifyStrategy":"BACKOFF_RETRY"}`
+
+var testTriggerMnsTopicTemplateUpdate = `{"filterTag":"anotherTag","notifyContentFormat":"JSON","notifyStrategy":"EXPONENTIAL_DECAY_RETRY"}`
+
+var testFCMnsTopicPolicyTemplate = `
+    {
+      "Version": "1",
+      "Statement": [
+        {
+          "Action": [
+            "log:PostLogStoreLogs"
+          ],
+          "Resource": "*",
+          "Effect": "Allow"
+        }
+      ]
+    }
+`
+
+var testFCMnsTopicRoleTemplate = `
+{
+  "Statement": [
+    {
+      "Action": "sts:AssumeRole",
+      "Effect": "Allow",
+      "Principal": {
+        "Service": [
+          "mns.aliyuncs.com"
+        ]
+      }
+    }
+  ],
+  "Version": "1"
+}
+`
+
 func testAccCheckAlicloudFCTriggerExists(name string, trigger *fc.GetTriggerOutput) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[name]
@@ -64,8 +326,7 @@ func testAccCheckAlicloudFCTriggerExists(name string, trigger *fc.GetTriggerOutp
 
 		client := testAccProvider.Meta().(*connectivity.AliyunClient)
 		fcService := FcService{client}
-		split := strings.Split(rs.Primary.ID, COLON_SEPARATED)
-		ser, err := fcService.DescribeFcTrigger(split[0], split[1], split[2])
+		ser, err := fcService.DescribeFcTrigger(rs.Primary.ID)
 		if err != nil {
 			return err
 		}
@@ -83,9 +344,7 @@ func testAccCheckAlicloudFCTriggerDestroy(s *terraform.State) error {
 		if rs.Type != "alicloud_fc_trigger" {
 			continue
 		}
-
-		split := strings.Split(rs.Primary.ID, COLON_SEPARATED)
-		ser, err := fcService.DescribeFcTrigger(split[0], split[1], split[2])
+		ser, err := fcService.DescribeFcTrigger(rs.Primary.ID)
 		if err != nil {
 			if NotFoundError(err) {
 				continue